Crate icu_provider_adapters

Adapters for composing and manipulating data providers.

Use the fork module to marshall data requests between multiple possible providers.

Use the either module to choose between multiple provider types at runtime.

Use the filter module to programmatically reject certain data requests.

Use the fallback module to automatically resolve arbitrary locales for data loading.

Modules

either

Helpers for switching between multiple providers.

empty

Empty data provider implementations.

fallback

A data provider wrapper that performs locale fallback.

filter

Providers that filter resource requests.

fixed

Data provider always serving the same struct.

fork

Providers that combine multiple other providers.