Struct icu_provider_adapters::any_payload::AnyPayloadProvider

pub struct AnyPayloadProvider {
    key: DataKey,
    data: AnyPayload,
}

A data provider that returns clones of a fixed type-erased payload.

AnyPayloadProvider implements AnyProvider, so it can be used in *_with_any_provider constructors across ICU4X.

Examples

use icu_provider::hello_world::*;
use icu_provider::prelude::*;
use icu_provider_adapters::any_payload::AnyPayloadProvider;
use std::borrow::Cow;
use writeable::assert_writeable_eq;

let provider =
    AnyPayloadProvider::from_static::<HelloWorldV1Marker>(&HelloWorldV1 {
        message: Cow::Borrowed("custom hello world"),
    });

// Check that it works:
let formatter = HelloWorldFormatter::try_new_with_any_provider(
    &provider,
    &Default::default(),
)
.expect("key matches");
assert_writeable_eq!(formatter.format(), "custom hello world");

// Requests for invalid keys get MissingDataKey
assert!(matches!(
    provider.load_any(icu_provider::data_key!("foo@1"), Default::default()),
    Err(DataError {
        kind: DataErrorKind::MissingDataKey,
        ..
    })
))

Fields
key: DataKey

The DataKey for which to provide data. All others will receive a DataErrorKind::MissingDataKey.

data: AnyPayload

The AnyPayload to return on matching requests.

Implementations

impl AnyPayloadProvider

pub fn from_owned<M: KeyedDataMarker>(data: M::Yokeable) -> Self
where M::Yokeable: MaybeSendSync,

Creates an AnyPayloadProvider with an owned (allocated) payload of the given data.

pub fn from_static<M: KeyedDataMarker>(data: &'static M::Yokeable) -> Self

Creates an AnyPayloadProvider with a statically borrowed payload of the given data.

pub fn from_payload<M: KeyedDataMarker>(payload: DataPayload<M>) -> Self
where M::Yokeable: MaybeSendSync,

Creates an AnyPayloadProvider from an existing DataPayload.

pub fn from_any_payload<M: KeyedDataMarker>(payload: AnyPayload) -> Self

Creates an AnyPayloadProvider from an existing AnyPayload.

pub fn new_default<M: KeyedDataMarker>() -> Self
where M::Yokeable: Default + MaybeSendSync,

Creates an AnyPayloadProvider with the default (allocated) version of the data struct.