Struct icu_provider_adapters::fallback::LocaleFallbacker

pub struct LocaleFallbacker { /* private fields */ }

Entry type for locale fallbacking.

See the module-level documentation for an example.

Implementations

impl LocaleFallbacker

pub fn try_new_unstable<P>(provider: &P) -> Result<Self, DataError>
where P: DataProvider<LocaleFallbackLikelySubtagsV1Marker> + DataProvider<LocaleFallbackParentsV1Marker> + DataProvider<CollationFallbackSupplementV1Marker> + ?Sized,

Creates a LocaleFallbacker with fallback data (likely subtags and parent locales).

๐ Help choosing a constructor

โ๏ธ The bounds on this function may change over time, including in SemVer minor releases.

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized) ) -> Result<Self, DataError>

Creates a new instance using an AnyProvider.

For details on the behavior of this function, see: Self::try_new_unstable

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized) ) -> Result<Self, DataError>

โจ Enabled with the "serde" feature.

Creates a new instance using a BufferProvider.

For details on the behavior of this function, see: Self::try_new_unstable

๐ Help choosing a constructor

pub fn new_without_data() -> Self

Creates a LocaleFallbacker without fallback data. Using this constructor may result in surprising behavior, especially in multi-script languages.

pub fn for_config( &self, config: LocaleFallbackConfig ) -> LocaleFallbackerWithConfig<'_>

Creates the intermediate LocaleFallbackerWithConfig with configuration options.

pub fn for_key(&self, data_key: DataKey) -> LocaleFallbackerWithConfig<'_>

Creates the intermediate LocaleFallbackerWithConfig based on a DataKey and a DataRequestMetadata.

Examples

use icu_locid::locale;
use icu_provider::prelude::*;
use icu_provider_adapters::fallback::LocaleFallbacker;
use std::borrow::Cow;

// Define the data struct with key.
#[icu_provider::data_struct(marker(
    FooV1Marker,
    "demo/foo@1",
    fallback_by = "region"
))]
pub struct FooV1<'data> {
    message: Cow<'data, str>,
};

// Set up the fallback iterator.
let fallbacker =
    LocaleFallbacker::try_new_unstable(&icu_testdata::unstable())
        .expect("data");
let key_fallbacker = fallbacker.for_key(FooV1Marker::KEY);
let mut fallback_iterator =
    key_fallbacker.fallback_for(DataLocale::from(locale!("en-GB")));

// Run the algorithm and check the results.
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("en-GB")));
fallback_iterator.step();
assert_eq!(
    fallback_iterator.get(),
    &DataLocale::from(locale!("und-GB"))
);
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("und")));