Struct icu_provider_adapters::fork::by_error::ForkByErrorProvider

pub struct ForkByErrorProvider<P0, P1, F>(P0, P1, F);

A provider that returns data from one of two child providers based on a predicate function.

This is an abstract forking provider that must be provided with a type implementing the ForkByErrorPredicate trait.

ForkByErrorProvider does not support forking between DataProviders. However, it supports forking between AnyProvider, BufferProvider, and DynamicDataProvider.

Tuple Fields
0: P01: P12: F
Implementations

impl<P0, P1, F> ForkByErrorProvider<P0, P1, F>

pub fn new_with_predicate(p0: P0, p1: P1, predicate: F) -> Self

Create a new provider that forks between the two children.

The predicate argument should be an instance of a struct implementing ForkByErrorPredicate.

pub fn inner(&self) -> (&P0, &P1)

Returns references to the inner providers.

pub fn inner_mut(&mut self) -> (&mut P0, &mut P1)

Returns mutable references to the inner providers.

pub fn into_inner(self) -> (P0, P1)

Returns ownership of the inner providers to the caller.

impl<P0, P1> ForkByErrorProvider<P0, P1, MissingDataKeyPredicate>

pub fn new(p0: P0, p1: P1) -> Self

A provider that returns data from one of two child providers based on the key.

See ForkByKeyProvider.