Struct icu_provider_adapters::fork::predicates::MissingLocalePredicate

pub struct MissingLocalePredicate;

A predicate that allows forking providers to search for a provider that supports a particular locale, based on whether it returns DataErrorKind::MissingLocale.

Examples

Configure a multi-language data provider pointing at two language packs:

use icu_provider_adapters::fork::ForkByErrorProvider;
use icu_provider_adapters::fork::predicates::MissingLocalePredicate;
use icu_provider_fs::FsDataProvider;
use icu_provider::prelude::*;
use icu_provider::hello_world::HelloWorldV1Marker;
use icu_locid::locale;

// The `tests` directory contains two separate "language packs" for Hello World data.
let base_dir = std::path::PathBuf::from(std::env!("CARGO_MANIFEST_DIR"))
    .join("tests/data/langtest");
let provider_de = FsDataProvider::try_new(base_dir.join("de")).unwrap();
let provider_ro = FsDataProvider::try_new(base_dir.join("ro")).unwrap();

// Create the forking provider:
let provider = ForkByErrorProvider::new_with_predicate(
    provider_de,
    provider_ro,
    MissingLocalePredicate
);

// Test that we can load both "de" and "ro" data:

let german_hello_world: DataPayload<HelloWorldV1Marker> = provider
    .as_deserializing()
    .load(DataRequest {
        locale: &locale!("de").into(),
        metadata: Default::default(),
    })
    .expect("Loading should succeed")
    .take_payload()
    .expect("Data should be present");

assert_eq!("Hallo Welt", german_hello_world.get().message);

let romanian_hello_world: DataPayload<HelloWorldV1Marker> = provider
    .as_deserializing()
    .load(DataRequest {
        locale: &locale!("ro").into(),
        metadata: Default::default(),
    })
    .expect("Loading should succeed")
    .take_payload()
    .expect("Data should be present");

assert_eq!("Salut, lume", romanian_hello_world.get().message);

// We should not be able to load "en" data because it is not in the provider:

DataProvider::<HelloWorldV1Marker>::load(
    &provider.as_deserializing(),
    DataRequest {
        locale: &locale!("en").into(),
        metadata: Default::default(),
    }
)
.expect_err("No English data");