icu_provider_adapters::fork

Type Alias ForkByMarkerProvider

pub type ForkByMarkerProvider<P0, P1> = ForkByErrorProvider<P0, P1, MarkerNotFoundPredicate>;

Create a provider that returns data from one of two child providers based on the marker.

The result of the first provider that supports a particular DataMarkerInfo will be returned, even if the request failed for other reasons (such as an unsupported language). Therefore, you should add child providers that support disjoint sets of markers.

ForkByMarkerProvider does not support forking between DataProviders. However, it supports forking between BufferProvider, and DynamicDataProvider.

Examples

Normal usage:

use icu_locale::langid;
use icu_provider::hello_world::*;
use icu_provider::prelude::*;
use icu_provider_adapters::fork::ForkByMarkerProvider;

struct DummyBufferProvider;
impl DynamicDataProvider<BufferMarker> for DummyBufferProvider {
    fn load_data(
        &self,
        marker: DataMarkerInfo,
        req: DataRequest,
    ) -> Result<DataResponse<BufferMarker>, DataError> {
        Err(DataErrorKind::MarkerNotFound.with_req(marker, req))
    }
}

let forking_provider = ForkByMarkerProvider::new(
    DummyBufferProvider,
    HelloWorldProvider.into_json_provider(),
);

let provider = forking_provider.as_deserializing();

let german_hello_world: DataResponse<HelloWorldV1> = provider
    .load(DataRequest {
        id: DataIdentifierBorrowed::for_locale(&langid!("de").into()),
        ..Default::default()
    })
    .expect("Loading should succeed");

assert_eq!("Hallo Welt", german_hello_world.payload.get().message);

Stops at the first provider supporting a marker, even if the locale is not supported:

use icu_locale::{subtags::language, langid};
use icu_provider::hello_world::*;
use icu_provider::prelude::*;
use icu_provider_adapters::filter::FilterDataProvider;
use icu_provider_adapters::fork::ForkByMarkerProvider;

let forking_provider = ForkByMarkerProvider::new(
    FilterDataProvider::new(
        HelloWorldProvider.into_json_provider(),
        "Chinese"
    )
    .with_filter(|id| id.locale.language == language!("zh")),
    FilterDataProvider::new(
        HelloWorldProvider.into_json_provider(),
        "German"
    )
    .with_filter(|id| id.locale.language == language!("de")),
);

let provider: &dyn DataProvider<HelloWorldV1> =
    &forking_provider.as_deserializing();

// Chinese is the first provider, so this succeeds
let chinese_hello_world = provider
    .load(DataRequest {
        id: DataIdentifierBorrowed::for_locale(&langid!("zh").into()),
        ..Default::default()
    })
    .expect("Loading should succeed");

assert_eq!("你好世界", chinese_hello_world.payload.get().message);

// German is shadowed by Chinese, so this fails
provider
    .load(DataRequest {
        id: DataIdentifierBorrowed::for_locale(&langid!("de").into()),
        ..Default::default()
    })
    .expect_err("Should stop at the first provider, even though the second has data");

Aliased Type

pub struct ForkByMarkerProvider<P0, P1>(/* private fields */);

Implementations

impl<P0, P1> ForkByMarkerProvider<P0, P1>

pub fn new(p0: P0, p1: P1) -> Self

A provider that returns data from one of two child providers based on the marker.

See ForkByMarkerProvider.