Crate icu_provider_adapters

Adapters for composing and manipulating data providers.

Use the fork module to marshall data requests between multiple possible providers.

Use the either module to choose between multiple provider types at runtime.

Use the filter module to programmatically reject certain data requests.

Modules

either

Helpers for switching between multiple providers.

filter

Providers that filter resource requests.

fork

Providers that combine multiple other providers.

struct_provider

Data provider always serving the same struct.

Macros

make_forking_provider

Make a forking data provider with an arbitrary number of inner providers that are known at build time.