Module icu_provider_adapters::fallback

Tools for locale fallback, enabling arbitrary input locales to be mapped into the nearest locale with data.

The algorithm implemented in this module is called Flexible Vertical Fallback. Watch #2243 to track improvements to this algorithm and steps to enshrine the algorithm in CLDR.

Examples

Run the locale fallback algorithm:

use icu_locid::locale;
use icu_provider_adapters::fallback::LocaleFallbacker;
use icu_provider::prelude::*;

// Set up a LocaleFallbacker with data.
let fallbacker = LocaleFallbacker::try_new_unstable(&icu_testdata::unstable()).expect("data");

// Create a LocaleFallbackerWithConfig with a configuration for a specific key.
// By default, uses language priority with no additional extension keywords.
let key_fallbacker = fallbacker.for_config(Default::default());

// Set up the fallback iterator.
let mut fallback_iterator = key_fallbacker.fallback_for(DataLocale::from(locale!("hi-Latn-IN")));

// Run the algorithm and check the results.
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("hi-Latn-IN")));
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("hi-Latn")));
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("en-IN")));
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("en-001")));
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("en")));
fallback_iterator.step();
assert_eq!(fallback_iterator.get(), &DataLocale::from(locale!("und")));

Modules

provider

Data provider struct definitions for vertical fallback.

Structs

LocaleFallbackConfig

Configuration settings for a particular fallback operation.

LocaleFallbackIterator

Iteration type for locale fallback operations.

LocaleFallbackProvider

A data provider wrapper that performs locale fallback. This enables arbitrary locales to be handled at runtime.

LocaleFallbacker

Entry type for locale fallbacking.

LocaleFallbackerWithConfig

Intermediate type for spawning locale fallback iterators based on a specific configuration.