icu_provider_adapters

Module fork

Providers that combine multiple other providers.

Types of Forking Providers

Marker-Based

To fork between providers that support different data markers, see:

ForkByMarkerProvider

MultiForkByMarkerProvider

Locale-Based

To fork between providers that support different locales, see:

ForkByErrorProvider<IdentiferNotFoundPredicate>

MultiForkByErrorProvider<IdentiferNotFoundPredicate>

Examples

See:

ForkByMarkerProvider

MultiForkByMarkerProvider

IdentiferNotFoundPredicate

Modules

predicates

Collection of predicate traits and functions for forking providers.

Structs

ForkByErrorProvider

A provider that returns data from one of two child providers based on a predicate function.

MultiForkByErrorProvider

A provider that returns data from the first child provider passing a predicate function.

Type Aliases

ForkByMarkerProvider

Create a provider that returns data from one of two child providers based on the marker.

MultiForkByMarkerProvider

A provider that returns data from the first child provider supporting the marker.