Module icu_provider_adapters::fork

Providers that combine multiple other providers.

Types of Forking Providers

Key-Based

To fork between providers that support different data keys, see:

ForkByKeyProvider

MultiForkByKeyProvider

Locale-Based

To fork between providers that support different locales, see:

ForkByErrorProvider<MissingLocalePredicate>

MultiForkByErrorProvider<MissingLocalePredicate>

Examples

See:

ForkByKeyProvider

MultiForkByKeyProvider

MissingLocalePredicate

Modules

by_error 🔒

macros 🔒

predicates

Collection of predicate traits and functions for forking providers.

Structs

ForkByErrorProvider

A provider that returns data from one of two child providers based on a predicate function.

MultiForkByErrorProvider

A provider that returns data from the first child provider passing a predicate function.

Type Aliases

ForkByKeyProvider

Create a provider that returns data from one of two child providers based on the key.

MultiForkByKeyProvider

A provider that returns data from the first child provider supporting the key.