Struct icu_provider_adapters::fork::by_error::MultiForkByErrorProvider

pub struct MultiForkByErrorProvider<P, F> {
    providers: Vec<P>,
    predicate: F,
}

A provider that returns data from the first child provider passing a predicate function.

This is an abstract forking provider that must be provided with a type implementing the ForkByErrorPredicate trait.

MultiForkByErrorProvider does not support forking between DataProviders. However, it supports forking between AnyProvider, BufferProvider, and DynamicDataProvider.

Fields
providers: Vec<P>predicate: F
Implementations

impl<P, F> MultiForkByErrorProvider<P, F>

pub fn new_with_predicate(providers: Vec<P>, predicate: F) -> Self

Create a new provider that forks between the vector of children.

The predicate argument should be an instance of a struct implementing ForkByErrorPredicate.

pub fn inner(&self) -> &[P]

Returns a slice of the inner providers.

pub fn with_inner_mut(&mut self, f: impl FnOnce(&mut Vec<P>))

Exposes a mutable vector of providers to a closure so it can be mutated.

pub fn into_inner(self) -> Vec<P>

Returns ownership of the inner providers to the caller.

pub fn push(&mut self, provider: P)

Adds an additional child provider.

impl<P> MultiForkByErrorProvider<P, MissingDataKeyPredicate>

pub fn new(providers: Vec<P>) -> Self

Create a provider that returns data from the first child provider supporting the key.

See MultiForkByKeyProvider.