Struct icu_provider_adapters::fork::MultiForkByErrorProvider

pub struct MultiForkByErrorProvider<P, F> { /* private fields */ }

A provider that returns data from the first child provider passing a predicate function.

This is an abstract forking provider that must be provided with a type implementing the ForkByErrorPredicate trait.

MultiForkByErrorProvider does not support forking between DataProviders. However, it supports forking between AnyProvider, BufferProvider, and DynamicDataProvider.

Implementations

impl<P, F> MultiForkByErrorProvider<P, F>

pub fn new_with_predicate(providers: Vec<P>, predicate: F) -> Self

Create a new provider that forks between the vector of children.

The predicate argument should be an instance of a struct implementing ForkByErrorPredicate.

pub fn inner(&self) -> &[P]

Returns a slice of the inner providers.

pub fn into_inner(self) -> Vec<P>

Returns ownership of the inner providers to the caller.

impl<P> MultiForkByErrorProvider<P, MissingDataKeyPredicate>

pub fn new(providers: Vec<P>) -> Self

Create a provider that returns data from the first child provider supporting the key.

See MultiForkByKeyProvider.