Type Alias icu_provider_adapters::fork::MultiForkByKeyProvider

pub type MultiForkByKeyProvider<P> = MultiForkByErrorProvider<P, MissingDataKeyPredicate>;

A provider that returns data from the first child provider supporting the key.

The result of the first provider that supports a particular DataKey will be returned, even if the request failed for other reasons (such as an unsupported language). Therefore, you should add child providers that support disjoint sets of keys.

MultiForkByKeyProvider does not support forking between DataProviders. However, it supports forking between AnyProvider, BufferProvider, and DynamicDataProvider.

Examples

use icu_locid::{subtags_language as language, locale};
use icu_provider::hello_world::*;
use icu_provider::prelude::*;
use icu_provider_adapters::filter::Filterable;
use icu_provider_adapters::fork::MultiForkByKeyProvider;

let forking_provider = MultiForkByKeyProvider::new(
    vec![
        HelloWorldProvider
            .into_json_provider()
            .filterable("Chinese")
            .filter_by_langid(|langid| langid.language == language!("zh")),
        HelloWorldProvider
            .into_json_provider()
            .filterable("German")
            .filter_by_langid(|langid| langid.language == language!("de")),
    ],
);

let data_provider: &dyn DataProvider<HelloWorldV1Marker> =
    &forking_provider.as_deserializing();

// Chinese is the first provider, so this succeeds
let chinese_hello_world = data_provider
    .load(DataRequest {
        locale: &locale!("zh").into(),
        metadata: Default::default(),
    })
    .expect("Loading should succeed")
    .take_payload()
    .expect("Data should be present");

assert_eq!("你好世界", chinese_hello_world.get().message);

// German is shadowed by Chinese, so this fails
data_provider
    .load(DataRequest {
        locale: &locale!("de").into(),
        metadata: Default::default(),
    })
    .expect_err("Should stop at the first provider, even though the second has data");

Aliased Type

struct MultiForkByKeyProvider<P> { /* private fields */ }

Implementations

impl<P> MultiForkByKeyProvider<P>

pub fn new(providers: Vec<P>) -> Self

Create a provider that returns data from the first child provider supporting the key.

See MultiForkByKeyProvider.

impl<P, F> MultiForkByErrorProvider<P, F>

pub fn new_with_predicate(providers: Vec<P>, predicate: F) -> Self

Create a new provider that forks between the vector of children.

The predicate argument should be an instance of a struct implementing ForkByErrorPredicate.

pub fn inner(&self) -> &[P]

Returns a slice of the inner providers.

pub fn with_inner_mut(&mut self, f: impl FnOnce(&mut Vec<P>))

Exposes a mutable vector of providers to a closure so it can be mutated.

pub fn into_inner(self) -> Vec<P>

Returns ownership of the inner providers to the caller.

pub fn push(&mut self, provider: P)

Adds an additional child provider.